UPSTREAM: mmc: sdhci: Check card status after reset
In sdhci_do_reset we call the reset callback which is typically
sdhci_reset. sdhci_reset can wait for up to 100ms waiting for the
controller to reset. If SDHCI_RESET_ALL was passed as the flag, the
controller will clear the IRQ mask. If during that 100ms the card is
removed there is no notification to the MMC system that the card was
removed. So from the drivers point of view the card is always present.

By making sdhci_reinit compare the present state it can schedule a
rescan if the card was removed while a reset was in progress.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Acked-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit dcaac3f7840177ec9894abbd4a0fabf31801ecb6)

BUG=b:206622593
TEST=Suspend/resume tests

Change-Id: Ie7fed47d1b300655864f6255618ff318bad0c598
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4817674
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>