UPSTREAM: drm/amd/display: Fix incorrect cursor pos on scaled primary plane
[Why]
Cursor pos is correctly adjusted from DC side for source rect offset
on DCN ASIC, but only on the overlay.

This is because DM places offsets the cursor for primary planes only
to workaround missing code in DCE for the adjustment we're now correctly
doing in DC for DCN ASIC.

[How]
Drop the adjustment for source rect from the DM side of things and put
the code where it actually belongs - in DC on the pipe level.

This matches what we do for DCN now.

Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: Zhan Liu <Zhan.Liu@amd.com>
Acked-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 03a4059b8dddf6a299ca941cc6e0d66bb146d884
 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

BUG=b:226417611
TEST=stop powerd && stop ui then plane_test --format NV12 --plane --format
     XR24 --size 100x100

Signed-off-by: wenchieh.chien <wenchieh.chien@amd.corp-partner.google.com>
Change-Id: I44ad5fc7da7796543dd700a62b7747cedccf19a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3818067
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>