BACKPORT: nl80211: reduce non-split wiphy dump size
When wiphy dumps cannot be split, such as in events or with
older userspace that doesn't support it, the size can today
be too big.

Reduce it, by doing two things:

 1) remove data that couldn't have been present before the
    split capability was introduced since it's new, such as
    HE capabilities

 2) as suggested by Martin Willi, remove management frame
    subtypes from the split dumps, as just (1) isn't even
    enough due to other new code capabilities. This is fine
    as old consumers (really just wpa_supplicant) didn't
    check this data before they got support for split dumps.

Reported-by: Martin Willi <martin@strongswan.org>
Suggested-by: Martin Willi <martin@strongswan.org>
Tested-by: Martin Willi <martin@strongswan.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Link: https://lore.kernel.org/r/20200928130655.53bce7873164.I71f06c9a221cd0630429a1a56eeae68a13beca61@changeid
(cherry picked from commit f8d504caa9739f74c7c399e158a7dce8478b866e)

Backport notes:
This change adds a boolean value "large" to indicate that an
nl80211 message is larger than a certain threshold. This
boolean value is used in several places, one of which is
related to channel frequency offsets,  which is not yet
implemented in 5.4. Because this feature does not exist in
5.4, and adding it in would require pulling a large number
of patches, it seemssafer to just remove that callsite from
this patch.

BUG=b:266832532
TEST=deploy to hatch device. Ensure that NL80211_CMD_NEW_WIPHY
are broadcast as expected and do basic test for WiFi
funtionality.
Signed-off-by: Kevin Lund <kglund@google.com>

Change-Id: I7e2cc3844adf71b363e4a333672eeb28e5a60f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4227499
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>