BACKPORT: cfg80211: Add support to configure SAE PWE value to drivers
Add support to configure SAE PWE preference from userspace to drivers in
both AP and STA modes. This is needed for cases where the driver takes
care of Authentication frame processing (SME in the driver) so that
correct enforcement of the acceptable PWE derivation mechanism can be
performed.

The userspace applications can pass the sae_pwe value using the
NL80211_ATTR_SAE_PWE attribute in the NL80211_CMD_CONNECT and
NL80211_CMD_START_AP commands to the driver. This allows selection
between the hunting-and-pecking loop and hash-to-element options for PWE
derivation. For backwards compatibility, this new attribute is optional
and if not included, the driver is notified of the value being
unspecified.

Signed-off-by: Rohan Dutta <drohan@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Link: https://lore.kernel.org/r/20201027100910.22283-1-jouni@codeaurora.org
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 9f0ffa418483938d25a15f6ad3891389f333bc59)
Signed-off-by: Jintao Lin <jintaolin@chromium.org>

Conflicts:
    include/uapi/linux/nl80211.h
    net/wireless/nl80211.c

BUG=b:259239952
TEST=build kernel
TEST=on Dedede, can start SoftAP in WPA3 and WPA3mixed

Change-Id: I22ff2939462165162197517ee19143aca5b866a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4749793
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>
Tested-by: Jintao Lin <jintaolin@chromium.org>