prevent inlining of __restore_processor_state
If we have a code that is compiled without a stack protector
(such as code in the Linux kernel that's trying to restore
registers containing the stack canaries), then inlining code
that was compiled with stack protector into it can break code.
LTO makes this more likely to occur across TUs which may or may
not be compiled with the same -fstack-protector flag.

Instead of disabling LTO to address issues like reset issue
on resume caused by above issue, prevent inlining of
__restore_processor_state function.

Tracked-On: OAM-96390
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>