UPSTREAM: media: mc: Provide a helper for setting bus_info field
The bus_info or a similar field exists in a lot of structs, yet drivers
tend to set the value of that field by themselves in a determinable way.
Thus provide a helper for doing this. To be used in subsequent patches.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit 78a171e58717685ea0eb082a24bc75bd9e20518f)

BUG=b:267802916
TEST=`media-ctl -p` on decoder should show bus_info

Change-Id: Ic41cd5803b6b97951f6ac8b1379c0a84a238c8f6
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4230617
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Steve Cho <stevecho@chromium.org>