CHROMIUM: dm-verity: Remove call to chromeos_set_need_recovery
chromeos_set_need_recovery would write the need to recover in eeprom.
It only works for ACPI based machines, and is not necessary as we also
write a prembule in kernel partition and coreboot is checking for the
string on all platforms.

BUG=b:172220950
TEST=check corrupting the root partition on eve-arc-r trigger image
invalidation.

Conflicts:
        drivers/md/dm-verity-chromeos.c: conflict with "use GPT try counter" recent changes.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: Id31a7eadc3f0adfc1f7656ca70bfb3103f60700b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3423086
Reviewed-by: Julius Werner <jwerner@chromium.org>
(cherry picked from commit c72335b4ac75042f095b6435067879fc460578d6)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4388811
Reviewed-by: Daniil Lunev <dlunev@chromium.org>