UPSTREAM: thunderbolt: Link USB4 ports to their USB Type-C connectors
Creating a symlink pointing to the correct USB Type-C
connector for the on-board USB4 ports when they are created.
The link will be created only if the firmware is able to
describe the connection between the port and its connector.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Won Chung <wonchung@google.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 5dddb41692849594862473e4ddf57fb52e1249d6)

BUG=b:197868988
TEST=built and deployed to a Drobit board, no issues booting.

Change-Id: I15eefb00ff90936dea687ddc30003755dc36140c
Signed-off-by: Jameson Thies <jthies@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4343157
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Benson Leung <bleung@google.com>