CHROMIUM: iwl7000: mvm: refactor __iwl_mvm_assign_vif_chanctx()
Since parts of the functionality of this function is going to be used
also by the MLD version of it, put in a separate function the parts
that are common for both MLD and non-MLD modes.
The common function will later be used in the MLD ops.

Change-Id: I1c650718ad2381eabc38f4103c1aac67936a1ffc
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 43b9bca0dc462a1de4851fd735750caa58ae3f34