CHROMIUM: iwl7000: mac80211: mlme: track assoc_bss/associated separately
We currently track whether we're associated and which the
BSS is in the same variable (ifmgd->associated), but for
MLD we'll need to move the BSS pointer to be per link,
while the question whether we're associated or not is for
the whole interface.

Add ifmgd->assoc_bss that stores the pointer and change
ifmgd->associated to be just a bool, so the question of
whether we're associated can continue working after MLD
rework, without requiring changes, while the BSS pointer
will have to be changed/used checked per link.

Change-Id: I1fb4635be57736eb525b7d41fa9a9f98105601bb
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 6aa6b9e979315fe2529b92927d5334853daecf6b