CHROMIUM: iwl7000: mvm: refactor iwl_mvm_sta_send_to_fw()
We have a new STA cmd as part of the new MLD API. There are some
parts of sending the STA cmd which are common to both the old and
the new one. Put this parts in functions which will later be used
to send the new STA cmd.

Change-Id: I3a66990fbffe9611b5e41f3686c2aff37ba2eb56
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 3d252b4dcdee14f82e3993d17d7f995754bf56d2