CHROMIUM: gpu: mali: Fix handling of remap or partial unmap for User IO pages mapping
GPUCORE-34473: Fix handling of remap or partial unmap for User IO pages mapping

This commit fixes the use after free and kernel panic issue that would
have happened if mremap() or partial munmap() is done on the mapping of
User IO pages.

For CSF GPUs, a memory mapping of 3 pages in size is created to allow
Userspace direct access to the CS_USER Input/Output pages and the HW
doorbell page. There is no use case to perform the mremap() or partial
munmap() on the IO pages mapping, so there was some code added in Kbase
added to fail such attempts made by a malicious Userspace. But the
handling was insufficient and that resulted in possibility of use after
free or kernel panic.

Kernel panic would have happened with mremap() if the 1st page (which
corresponds to the Doorbell page) of new mapping was accessed.
Use after free would have happened on the access to 2nd & 3rd page of
the new mapping.

With partial munmap(), use after free would have happened on the access
to 3rd page if the 2nd page was first unmappped first, followed by the
unmap of 1st page.

This commit ensures that both mremap() and partial munmap() calls made
for the User IO pages mapping itself will return an error.

It has been assumed that for CSF GPUs the minimum supported kernel
version is >= v4.9.67.

TI2: 869611 (PLAN-11965r693 DDK Precommit, build only jobs)
TI2: 869065 (Few scenarios from PLAN-12467r454 TGT CS Nightly)

BUG=b:247578925
TEST=tast run $IP graphics.DEQP \
                  graphics.GLBench.hasty \
                  graphics.VKBench
TEST=test_that $IP graphics_parallel_dEQP.gles31 \
                   graphics_WebGLAquarium.thermal

Change-Id: I809a0b1349daedc151b7c1598f1d669c57bc247a
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3944569
Reviewed-by: Wei-Shun Chang <weishunc@chromium.org>