UPSTREAM: r8152: fix a WOL issue
This fixes that the platform is waked by an unexpected packet. The
size and range of FIFO is different when the device enters S3 state,
so it is necessary to correct some settings when suspending.

Regardless of jumbo frame, set RMS to 1522 and MTPS to MTPS_DEFAULT.
Besides, enable MCU_BORW_EN to update the method of calculating the
pointer of data. Then, the hardware could get the correct data.

Fixes: 195aae321c82 ("r8152: support new chips")
Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Link: https://lore.kernel.org/r/20220718082120.10957-391-nic_swsd@realtek.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit cdf0b86b250fd3c1c3e120c86583ea510c52e4ce)

BUG=none
TEST=test_that -b brya  192.168.86.163 network_EthernetStressPlug --args num_iterations=1000
     using redrix (brya) + ServoV4.0 connected to netgear GS108 switch

Change-Id: Id77c64021a33e8e04dcc1c8ca1bfbd1416876688
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4129448
Reviewed-by: Hayes Wang <hayeswang@realtek.com>
Reviewed-by: Ying Hsu <yinghsu@chromium.org>