CHROMIUM: iwl7000: mac80211: consider reserved chanctx for mindef
When a chanctx is reserved for a new vif and we recalculate
the minimal definition for it, we need to consider the new
interface it's being reserved for before we assign it, so it
can be used directly with the correct min channel width.

Fix the code to - optionally - consider that, and use that
option just before doing the reassignment.

Change-Id: Icfa2f08e87dddb9f74199ba1bbef2e926f24fe2d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 309597beb404cd811bfca67ec00e65dd6fa6b2bf