BACKPORT: FROMGIT: drm/i915/pxp: Invalidate all PXP fw sessions during teardown
A gap was recently discovered where if an application did not
invalidate all of the stream keys (intentionally or not), and the
driver did a full PXP global teardown on the GT subsystem, we
find that future session creation would fail on the security
firmware's side of the equation. i915 is the entity that needs
ensure the sessions' state across both iGT and security firmware
are at a known clean point when performing a full global teardown.

That said, the i915 should inspect all active sessions and submit
the invalidate-stream-key PXP command to the security firmware for
each of them.

BUG=b:247565446
TEST=tast run <IP> video.PlayDRM*

Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Link:
https://patchwork.freedesktop.org/patch/520139/?series=113307&rev=1
(cherry picked from commit d374c047b38e9f1130308aae207dc44045cd5cac
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

Change-Id: I5b7591ac91c54754573c883c72945a252e926037
Tested-by: Lewis Chen <lewis.chen@intel.com>
Signed-off-by: Srihari Uttanur <srihari.uttanur@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4109081
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Juston Li <justonli@google.com>
Commit-Queue: Sam McNally <sammc@chromium.org>
Tested-by: Lewis Chen <lewis.chen@intel.corp-partner.google.com>