CHROMIUM: iwl7000: pcie: try to rebind/rescan for some devices
For "ma" devices that are integrated, some platforms have an
issue that causes the firmware to do some hardware reconfig
that causes the devices to fall off the bus.

If we detect that such a device is unresponsive, then assume
the NIC reset itself and try to disconnect from it and rescan
the bus when it happens.

Change-Id: I26ac973ae055b4902dde559990211dd45b4b33ee
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 88702da607a295e296cd8adf88d90139d2897818