CHROMIUM: dm-verity: use GPT try counter for integrity retries
This is to prevent a situation of transient errors which are not
detected by verity as transient - e.g. memory corruption, on the wire
flips or simillar. DM-verity will use the try counter in the associated
kernel partition (update-engine will be modified to set it properly) to
give the system a few reboot-retries before declaring the system
corrupted.

BUG=b:272211532
TEST=manual
UPSTREAM-TASK=b:149112468

Signed-off-by: Daniil Lunev <dlunev@chromium.org>

Change-Id: Ie385ec88e2111bfd5ba26ebf42ad9f0b60234d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4349543
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>