CHROMIUM: cfg80211: hold RTNL for auto-disconnect
If the userspace sets the control flag to have the connection
owned by the socket, then when userspace closes the socket or
crashes, the auto-disconnect work will run and disconnect the
connection (or leave IBSS, stop AP operation, etc.)

This disconnect work currently only holds the wdev mutex, but
all the others code paths that lead to a disconnect also hold
the RTNL.

Future related updates to iwl7000 require these to be aligned
as the RTNL will be used to exclude iwl7000_mac80211 workers;
this is needed to be from cfg80211 to avoid deadlocks there.

Modify the auto-disconnect work to be per wiphy/rdev so that
we can now hold the RTNL, and do that.

This is an adaptation of upstream commit e9da6df7492a ("wifi:
cfg80211: hold wiphy lock in auto-disconnect") which changed
the code there to hold the wiphy_lock() instead, but that was
only introduced in 5.12 with older kernels using the RTNL in
its place.

BUG=b:288121907
TEST=connect and killall -9 wpa_supplicant to invoke this path,
     with an assertion added to the driver that RTNL is held
     and a print to ensure deauth was called

Change-Id: I4f7c8453ba9524d410e2e43feb6bd338fffc95cd
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4632624
Reviewed-by: David Ruth <druth@chromium.org>
Commit-Queue: David Ruth <druth@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: David Ruth <druth@chromium.org>