CHROMIUM: cfg80211: make nl80211_attrs order align with the ones in the upstream kernel
The nl80211_attrs enum order does not match the one used in upstream
kernel when backported in the following commits,

 - commit 3b141f0e0704eb815de13b49ebc27a53bbb4b657
 - commit 0030a0654d0a5ed289e91a2a92522a2dd80d25a6
 - commit c3dd33d3f2bc23133abd4a56822a921613158329
 - commit 82fd35c11ca9a14851ac9e7bf8486e4371afded1

This mismatch causes troubles when userspace applications like
wpa_supplicant uses a nl80211_copy.h file which aligns with the upstream
kernel but not the one used in Chromium kernel. It will send mismatched
attibute values to kernel and cause error.

Instead of reverting and relanding the above commits and introduces
several conflicts in both reverting and relanding CLs which is prone to
errors. This CL is the lesser of two evils and fix the ordering in a
CHROMIUM patch.

Signed-off-by: Jintao Lin <jintaolin@chromium.org>

BUG=b:259239952
TEST=build kernel
TEST=can start mobile AP successfully

Change-Id: Ife7db2a6461e38d6b4a00f4bd27636a04e63df3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4743591
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>