UPSTREAM: wireguard: allowedips: expand maximum node depth
In the allowedips self-test, nodes are inserted into the tree, but it
generated an even amount of nodes, but for checking maximum node depth,
there is of course the root node, which makes the total number
necessarily odd. With two few nodes added, it never triggered the
maximum depth check like it should have. So, add 129 nodes instead of
128 nodes, and do so with a more straightforward scheme, starting with
all the bits set, and shifting over one each time. Then increase the
maximum depth to 129, and choose a better name for that variable to
make it clear that it represents depth as opposed to bits.

Cc: stable@vger.kernel.org
Fixes: e7096c131e51 ("net: WireGuard secure network tunnel")
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Link: https://lore.kernel.org/r/20230807132146.2191597-2-Jason@zx2c4.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 46622219aae2b67813fe31a7b8cb7da5baff5c8a)

BUG=b:202680072
TEST=`make -C tools/testing/selftests/wireguard/qemu -j$(nproc)`

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: I575207f26393a17288ce0c23f7ca642b4c4c197f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4771909
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Taoyu Li <taoyl@chromium.org>